refactor: promote Mirror to core/mirror.py
Second half of the substrate promotion. dataset/retrieval.py now holds only
the dataset resolution chain; the mirror is shared machinery.
