+++ defendable-science/tests/test_retrieval.py
@@ -224,63 +224,6 @@ def test_fetch_from_mirror(tmp_path: Path) -> None:
     assert paths[0].read_bytes() == payload
 
 
-# --- Mirror -----------------------------------------------------------------
-
-
-def test_mirror_check_and_target() -> None:
-    rclone = FakeRclone(present=True)
-    mirror = r.Mirror(
-        remote="store", base_path="proj", config_path="c.conf", run=rclone
-    )
-    assert mirror.check("sha256:" + "a" * 64) is True
-    assert rclone.calls[-1][:3] == ["rclone", "--config", "c.conf"]
-    assert rclone.calls[-1][-1] == "store:proj/sha256/" + "a" * 64
-
-
-def test_mirror_missing_binary_is_actionable() -> None:
-    def _no_binary(args: list[str], **_kw: object) -> FakeProc:
-        raise FileNotFoundError("rclone")
-
-    mirror = r.Mirror(remote="store", run=_no_binary)
-    with pytest.raises(r.RetrievalError, match="rclone not found"):
-        mirror.check("a" * 64)
-
-
-def test_mirror_forwards_scoped_env_merged_over_process_env(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    monkeypatch.setenv("PATH", "/usr/bin")
-    captured: dict[str, object] = {}
-
-    def _runner(args: list[str], **kw: object) -> FakeProc:
-        captured.update(kw)
-        return FakeProc(0)
-
-    mirror = r.Mirror(
-        remote="store",
-        run=_runner,
-        env={"RCLONE_CONFIG_STORE_TYPE": "s3"},
-    )
-    assert mirror.check("a" * 64) is True
-    passed = captured["env"]
-    assert isinstance(passed, dict)
-    # scoped secret is present …
-    assert passed["RCLONE_CONFIG_STORE_TYPE"] == "s3"
-    # … merged over the inherited environment (PATH survives).
-    assert passed["PATH"] == "/usr/bin"
-
-
-def test_mirror_without_env_passes_no_env_kwarg() -> None:
-    captured: dict[str, object] = {}
-
-    def _runner(args: list[str], **kw: object) -> FakeProc:
-        captured.update(kw)
-        return FakeProc(0)
-
-    r.Mirror(remote="store", run=_runner).check("a" * 64)
-    assert "env" not in captured
-
-
 # --- audit ------------------------------------------------------------------
 
 
@@ -336,12 +279,3 @@ def test_audit_with_mirror_present(tmp_path: Path) -> None:
         m.Manifest(datasets=[entry]), cache_dir=tmp_path / "c", mirror=mirror
     )
     assert report.mirror_present["d1"] is True
-
-
-def test_mirror_put_failure_raises() -> None:
-    def _fail(args: list[str], **_kw: object) -> FakeProc:
-        return FakeProc(1)
-
-    mirror = r.Mirror(remote="s", run=_fail)
-    with pytest.raises(r.RetrievalError, match="copyto to mirror failed"):
-        mirror.put("/tmp/x", "a" * 64)
